// apps/web/src/api/generated/models/widget/widget.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Talqo API
 * OpenAPI spec version: 0.0.1
 */
import * as zod from "zod"

import { WidgetAppearance } from "./widgetAppearance.zod"

export const Widget = zod.object({
	id: zod.string(),
	agentId: zod.string(),
	name: zod.string(),
	publicToken: zod.string(),
	appearance: WidgetAppearance,
})

export type Widget = zod.input<typeof Widget>
export type WidgetOutput = zod.output<typeof Widget>